JDBC backend

This backend supports databases accessed via JDBCConnection created by RJDBC::JDBC() and DBI::dbConnect(). dbplyr automatically detects the underlying database type from the JDBC connection class and uses the appropriate SQL dialect.

The following databases are recognized via their JDBC connection class:

PostgreSQL (org.postgresql.*)

MySQL (com.mysql.*)

MariaDB (org.mariadb.*)

SQLite (org.sqlite.*)

Oracle (oracle.*)

SQL Server (com.microsoft.sqlserver.*)

SAP HANA (com.sap.db.*)

Teradata (com.teradata.*)

Apache Hive (org.apache.hive.*)

Apache Spark (org.apache.spark.* or com.simba.spark.*)

Snowflake (net.snowflake.*)

Impala (com.cloudera.impala.*)

Amazon Redshift (com.amazon.redshift.*)

If your database is not recognized, dbplyr will fall back to a generic ODBC dialect. Please file an issue if you'd like support for additional databases.

See vignette("translation-function") and vignette("translation-verb") for details of overall translation technology.

Examples

# JDBC connections require the RJDBC package and a JDBC driver JAR file.
# Once connected, dbplyr automatically detects the database type:
#
# library(RJDBC)
# library(dplyr, warn.conflicts = FALSE)
#
# drv <- JDBC("org.postgresql.Driver", "postgresql.jar")
# con <- dbConnect(drv, "jdbc:postgresql://localhost/mydb", "user", "password")
# tbl(con, "my_table") |> filter(x > 1)