dbplyr-package {dbplyr}	R Documentation

dbplyr: A 'dplyr' Back End for Databases

Description

A 'dplyr' back end for databases that allows you to work with remote database tables as if they are in-memory data frames. Basic features works with any database that has a 'DBI' back end; more advanced features require 'SQL' translation to be provided by the package author.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Authors:

Edgar Ruiz

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://github.com/tidyverse/dbplyr

Report bugs at https://github.com/tidyverse/dplyr/issues

[Package dbplyr version 1.2.2 Index]