filter.tbl_lazy {dbplyr}	R Documentation

Subset rows using column values

Description

This is a method for the dplyr filter() generic. It generates the WHERE clause of the SQL query.

Usage


## S3 method for class 'tbl_lazy'
filter(.data, ..., .preserve = FALSE)

Arguments

.data
A lazy data frame backed by a database query.

...
<data-masking> Variables, or functions of variables. Use desc() to sort a variable in descending order.

.preserve
Not supported by this method.

Value

Another tbl_lazy. Use show_query() to see the generated query, and use collect() to execute the query and return data to R.

Examples


library(dplyr, warn.conflicts = FALSE)

db <- memdb_frame(x = c(2, NA, 5, NA, 10), y = 1:5)
db %>% filter(x < 5) %>% show_query()
db %>% filter(is.na(x)) %>% show_query()

[Package dbplyr version 2.1.1 Index]