ident {dbplyr}	R Documentation

Flag a character vector as SQL identifiers

Description

ident() takes strings and turns them as database identifiers (e.g. table or column names) quoting them using the identifer rules for your database. ident_q() does the same, but assumes the names have already been quoted, preventing them from being quoted again.

These are generally for internal use only; if you need to supply an table name that is qualified with schema or catalog, or has already been quoted for some other reason, use I().

Usage

ident(...)

is.ident(x)

Arguments

...
A character vector, or name-value pairs.

x
An object.

Examples

# SQL92 quotes strings with '
escape_ansi("x")

# And identifiers with "
ident("x")
escape_ansi(ident("x"))

# You can supply multiple inputs
ident(a = "x", b = "y")
ident_q(a = "x", b = "y")

[Package dbplyr version 2.5.0 Index]