lazy_ops {dbplyr}	R Documentation

Lazy operations

Description

This set of S3 classes describe the action of dplyr verbs. These are currently used for SQL sources to separate the description of operations in R from their computation in SQL. This API is very new so is likely to evolve in the future.

Usage

op_base(x, vars, class = character())

op_single(name, x, dots = list(), args = list())

add_op_single(name, .data, dots = list(), args = list())

op_double(name, x, y, args = list())

op_grps(op)

op_vars(op)

op_sort(op)

op_frame(op)

Details

op_vars() and op_grps() compute the variables and groups from a sequence of lazy operations. op_sort() tracks the order of the data for use in window functions.

[Package dbplyr version 1.2.2 Index]