mutate.tbl_lazy {dbplyr}	R Documentation

Create, modify, and delete columns

Description

These are methods for the dplyr mutate() and transmute() generics. They are translated to computed expressions in the SELECT clause of the SQL query.

Usage


## S3 method for class 'tbl_lazy'
mutate(.data, ...)

Arguments

.data
A lazy data frame backed by a database query.

...
<data-masking> Variables, or functions of variables. Use desc() to sort a variable in descending order.

Value

Another tbl_lazy. Use show_query() to see the generated query, and use collect() to execute the query and return data to R.

Examples


library(dplyr, warn.conflicts = FALSE)

db <- memdb_frame(x = 1:5, y = 5:1)
db %>%
  mutate(a = (x + y) / 2, b = sqrt(x^2L + y^2L)) %>%
  show_query()

# dbplyr automatically creates subqueries as needed
db %>%
  mutate(x1 = x + 1, x2 = x1 * 2) %>%
  show_query()

[Package dbplyr version 2.1.1 Index]