pull.tbl_sql {dbplyr}	R Documentation

Extract a single column

Description

This is a method for the dplyr pull() generic. It evaluates the query retrieving just the specified column.

Usage


## S3 method for class 'tbl_sql'
pull(.data, var = -1)

Arguments

.data
A lazy data frame backed by a database query.

var
A variable specified as:

a literal variable name

a positive integer, giving the position counting from the left

a negative integer, giving the position counting from the right.

The default returns the last column (on the assumption that's the column you've created most recently).

This argument is taken by expression and supports quasiquotation (you can unquote column names and column locations).

Value

A vector of data.

Examples


library(dplyr, warn.conflicts = FALSE)

db <- memdb_frame(x = 1:5, y = 5:1)
db %>%
  mutate(z = x + y * 2) %>%
  pull()

[Package dbplyr version 2.1.1 Index]