sql {dbplyr}	R Documentation

SQL escaping.

Description

These functions are critical when writing functions that translate R functions to sql functions. Typically a conversion function should escape all its inputs and return an sql object.

Usage

sql(...)

is.sql(x)

as.sql(x)

Arguments

...
Character vectors that will be combined into a single SQL expression.

x
Object to coerce

[Package dbplyr version 1.2.2 Index]