tbl_sql {dbplyr}	R Documentation

Create an SQL tbl (abstract)

Description

Generally, you should no longer need to provide a custom tbl() method you you can default tbl.DBIConnect method.

Usage

tbl_sql(subclass, src, from, ..., vars = NULL)

Arguments

subclass
name of subclass

...
needed for agreement with generic. Not otherwise used.

vars
If known, the names of the variables in the tbl. This is relatively expensive to determine automatically, so is cached throughout dplyr. However, you should usually be able to leave this blank and it will be determined from the context.

[Package dbplyr version 1.2.2 Index]