Build a SQL string.

build_sql() is superseded in favor of sql_glue2().

This is a convenience function that should prevent sql injection attacks (which in the context of dplyr are most likely to be accidental not deliberate) by automatically escaping all expressions in the input, while treating bare strings as sql. This is unlikely to prevent any serious attack, but should make it unlikely that you produce invalid sql.

This function should be used only when generating SELECT clauses, other high level queries, or for other syntax that has no R equivalent. For individual function translations, prefer sql_expr().

Usage

build_sql(..., .env = parent.frame(), con = sql_current_con())

Arguments

...

input to convert to SQL. Use sql() to preserve user input as is (dangerous), and ident() to label user input as sql identifiers (safe)

.env

the environment in which to evaluate the arguments. Should not be needed in typical use.

con

database connection; used to select correct quoting characters.

Examples

con <- dialect_ansi()

# Old:
build_sql("SELECT * FROM ", ident("table"), con = con)
#> <SQL> SELECT * FROM "table"
# New:
sql_glue2(con, "SELECT * FROM {.tbl 'table'}")
#> <SQL> SELECT * FROM "table"

# Old:
name <- "Robert"
build_sql("INSERT INTO students (name) VALUES (", name, ")", con = con)
#> <SQL> INSERT INTO students (name) VALUES ('Robert')
# New:
sql_glue2(con, "INSERT INTO students (name) VALUES ({name})")
#> <SQL> INSERT INTO students (name) VALUES ('Robert')