collapse.tbl_sql {dbplyr}	R Documentation

Force computation of query

Description

collapse() creates a subquery; compute() stores the results in a remote table; collect() downloads the results into the current R session.

Usage

## S3 method for class 'tbl_sql'
collapse(x, ...)

## S3 method for class 'tbl_sql'
compute(
  x,
  name = unique_table_name(),
  temporary = TRUE,
  unique_indexes = list(),
  indexes = list(),
  analyze = TRUE,
  ...
)

## S3 method for class 'tbl_sql'
collect(x, ..., n = Inf, warn_incomplete = TRUE)

Arguments

x
A tbl_sql

...
other parameters passed to methods.

name
Table name in remote database.

temporary
Should the table be temporary (TRUE, the default) or persistent (FALSE')?

unique_indexes
a list of character vectors. Each element of the list will create a new unique index over the specified column(s). Duplicate rows will result in failure.

indexes
a list of character vectors. Each element of the list will create a new index.

analyze
if TRUE (the default), will automatically ANALYZE the new table so that the query optimiser has useful information.

n
Number of rows to fetch. Defaults to Inf, meaning all rows.

warn_incomplete
Warn if n is less than the number of result rows?

[Package dbplyr version 1.4.4 Index]