db-misc {dbplyr}	R Documentation

Miscellaneous database generics

Description

db_connection_describe() provides a short string describing the database connection, helping users tell which database a table comes from. It should be a single line, and ideally less than 60 characters wide.

Usage

db_connection_describe(con)

sql_join_suffix(con, ...)

db_sql_render(con, sql, ..., cte = FALSE)

dbplyr_edition(con)

Details

dbplyr_edition() declares which version of the dbplyr API you want. See below for more details.

dbplyr 2.0.0

dbplyr 2.0.0 renamed a number of generics so that they could be cleanly moved from dplyr to dbplyr. If you have an existing backend, you'll need to rename the following methods.

dplyr::db_desc() -> dbplyr::db_connection_describe() (also note that the argument named changed from x to con).

See Also

Other generic: db-sql, db_copy_to(), sql_escape_logical()

[Package dbplyr version 2.3.4 Index]