dbplyr-package {dbplyr}	R Documentation

dbplyr: A 'dplyr' Back End for Databases

Description

A 'dplyr' back end for databases that allows you to work with remote database tables as if they are in-memory data frames. Basic features works with any database that has a 'DBI' back end; more advanced features require 'SQL' translation to be provided by the package author.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Authors:

Maximilian Girlich

Edgar Ruiz

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://dbplyr.tidyverse.org/

https://github.com/tidyverse/dbplyr

Report bugs at https://github.com/tidyverse/dbplyr/issues

[Package dbplyr version 2.3.4 Index]