dbplyr_uncount {dbplyr}	R Documentation

"Uncount" a database table

Description

This is a method for the tidyr uncount() generic. It uses a temporary table, so your database user needs permissions to create one.

Usage

dbplyr_uncount(data, weights, .remove = TRUE, .id = NULL)

Arguments

data
A lazy data frame backed by a database query.

weights
A vector of weights. Evaluated in the context of data; supports quasiquotation.

.remove
If TRUE, and weights is the name of a column in data, then this column is removed.

.id
Supply a string to create a new variable which gives a unique identifier for each created row.

Examples

df <- memdb_frame(x = c("a", "b"), n = c(1, 2))
dbplyr_uncount(df, n)
dbplyr_uncount(df, n, .id = "id")

# You can also use constants
dbplyr_uncount(df, 2)

# Or expressions
dbplyr_uncount(df, 2 / n)

[Package dbplyr version 2.3.4 Index]