Cache and retrieve an src_sqlite of the Lahman baseball database.

This creates an interesting database using data from the Lahman baseball data source, provided by Sean Lahman at http://www.seanlahman.com/baseball-archive/statistics/, and made easily available in R through the Lahman package by Michael Friendly, Dennis Murphy and Martin Monkman. See the documentation for that package for documentation of the inidividual tables.

lahman_sqlite(path = NULL)

lahman_postgres(dbname = "lahman", host = "localhost", ...)

lahman_mysql(dbname = "lahman", ...)

lahman_df()

copy_lahman(src, ...)

has_lahman(type, ...)

lahman_srcs(..., quiet = NULL)

Arguments

...
Other arguments passed to src on first load. For mysql and postgresql, the defaults assume you have a local server with lahman database already created. For lahman_srcs(), character vector of names giving srcs to generate.

type
src type.

quiet
if TRUE, suppress messages about databases failing to connect.

Examples

# Connect to a local sqlite database, if already created

if (has_lahman("sqlite")) {
  lahman_sqlite()
  batting <- tbl(lahman_sqlite(), "Batting")
  batting
}

#> Creating table: AllstarFull

#> Creating table: Appearances

#> Creating table: AwardsManagers

#> Creating table: AwardsPlayers

#> Creating table: AwardsShareManagers

#> Creating table: AwardsSharePlayers

#> Creating table: Batting

#> Creating table: BattingPost

#> Creating table: CollegePlaying

#> Creating table: Fielding

#> Creating table: FieldingOF

#> Creating table: FieldingPost

#> Creating table: HallOfFame

#> Creating table: LahmanData

#> Creating table: Managers

#> Creating table: ManagersHalf

#> Creating table: Master

#> Creating table: Pitching

#> Creating table: PitchingPost

#> Creating table: Salaries

#> Creating table: Schools

#> Creating table: SeriesPost

#> Creating table: Teams

#> Creating table: TeamsFranchises

#> Creating table: TeamsHalf

#> Source:     table<Batting> [?? x 22]
#> Database:   sqlite 3.11.1 [/Users/hadley/R/dplyr/db/lahman.sqlite]
#> 
#> # S3: tbl_dbi
#>    playerID yearID stint teamID  lgID     G    AB     R     H   X2B   X3B    HR
#>       <chr>  <int> <int>  <chr> <chr> <int> <int> <int> <int> <int> <int> <int>
#> 1 abercda01   1871     1    TRO    NA     1     4     0     0     0     0     0
#> 2  addybo01   1871     1    RC1    NA    25   118    30    32     6     0     0
#> 3 allisar01   1871     1    CL1    NA    29   137    28    40     4     5     0
#> 4 allisdo01   1871     1    WS3    NA    27   133    28    44    10     2     2
#> 5 ansonca01   1871     1    RC1    NA    25   120    29    39    11     3     0
#> 6 armstbo01   1871     1    FW1    NA    12    49     9    11     2     1     0
#> # ... with more rows, and 10 more variables: RBI <int>, SB <int>, CS <int>,
#> #   BB <int>, SO <int>, IBB <int>, HBP <int>, SH <int>, SF <int>, GIDP <int>

# Connect to a local postgres database with lahman database, if available
if (has_lahman("postgres")) {
  lahman_postgres()
  batting <- tbl(lahman_postgres(), "Batting")
}