memdb_frame {dbplyr}	R Documentation

Create a database table in temporary in-memory database.

Description

memdb_frame() works like tibble::tibble(), but instead of creating a new data frame in R, it creates a table in src_memdb().

Usage

memdb_frame(..., .name = random_table_name())

src_memdb()

Arguments

...
A set of name-value pairs. Arguments are evaluated sequentially, so you can refer to previously created variables. These arguments are processed with rlang::quos() and support unquote via !! and unquote-splice via !!!.

.name
Name of table in database: defaults to a random name that's unlikely to conflict with an existing table.

Examples

library(dplyr)
df <- memdb_frame(x = runif(100), y = runif(100))
df %>% arrange(x)
df %>% arrange(x) %>% show_query()

[Package dbplyr version 1.2.2 Index]