Create a database table in temporary in-memory database.

memdb_frame() works like tibble::tibble(), but instead of creating a new data frame in R, it creates a table in src_memdb().

memdb_frame(..., .name = unique_table_name())

tbl_memdb(df, name = deparse(substitute(df)))

src_memdb()

Arguments

...
<dynamic-dots> A set of name-value pairs. These arguments are processed with rlang::quos() and support unquote via !! and unquote-splice via !!!. Use := to create columns that start with a dot.

Arguments are evaluated sequentially. You can refer to previously created elements directly or using the .data pronoun. An existing .data pronoun, provided e.g. inside dplyr::mutate(), is not available.

df
Data frame to copy

name, .name
Name of table in database: defaults to a random name that's unlikely to conflict with an existing table.

Examples

library(dplyr)
df <- memdb_frame(x = runif(100), y = runif(100))
df %>% arrange(x)

#> # Source:     table<dbplyr_016> [?? x 2]
#> # Database:   sqlite 3.34.0 [:memory:]
#> # Ordered by: x
#>          x     y
#>      <dbl> <dbl>
#>  1 0.00707 0.846
#>  2 0.0299  0.181
#>  3 0.0436  0.149
#>  4 0.0649  0.161
#>  5 0.0682  0.679
#>  6 0.0753  0.154
#>  7 0.0776  0.993
#>  8 0.0801  0.663
#>  9 0.0998  0.961
#> 10 0.111   0.709
#> # … with more rows

df %>% arrange(x) %>% show_query()

#> <SQL>
#> SELECT *
#> FROM `dbplyr_016`
#> ORDER BY `x`

mtcars_db <- tbl_memdb(mtcars)
mtcars_db %>% group_by(cyl) %>% summarise(n = n()) %>% show_query()

#> <SQL>
#> SELECT `cyl`, COUNT(*) AS `n`
#> FROM `mtcars`
#> GROUP BY `cyl`