Create a database table in temporary in-memory database.

memdb_frame() works like tibble::tibble(), but instead of creating a new data frame in R, it creates a table in src_memdb().

Usage

memdb_frame(..., .name = unique_table_name())

tbl_memdb(df, name = deparse(substitute(df)))

src_memdb()

Arguments

...

<dynamic-dots> A set of name-value pairs. These arguments are processed with rlang::quos() and support unquote via !! and unquote-splice via !!!. Use := to create columns that start with a dot.

Arguments are evaluated sequentially. You can refer to previously created elements directly or using the .data pronoun. To refer explicitly to objects in the calling environment, use !! or .env, e.g. !!.data or .env$.data for the special case of an object named .data.

df

Data frame to copy

name, .name

Name of table in database: defaults to a random name that's unlikely to conflict with an existing table.

Examples

library(dplyr)
df <- memdb_frame(x = runif(100), y = runif(100))
df %>% arrange(x)
#> # Source:     SQL [?? x 2]
#> # Database:   sqlite 3.40.1 [:memory:]
#> # Ordered by: x
#>          x      y
#>      <dbl>  <dbl>
#>  1 0.00905 0.0618
#>  2 0.0373  0.632 
#>  3 0.0388  0.823 
#>  4 0.0408  0.897 
#>  5 0.0471  0.725 
#>  6 0.0669  0.328 
#>  7 0.0711  0.0298
#>  8 0.0842  0.0777
#>  9 0.111   0.450 
#> 10 0.115   0.906 
#> # … with more rows
df %>% arrange(x) %>% show_query()
#> <SQL>
#> SELECT *
#> FROM `dbplyr_018`
#> ORDER BY `x`

mtcars_db <- tbl_memdb(mtcars)
mtcars_db %>% group_by(cyl) %>% summarise(n = n()) %>% show_query()
#> <SQL>
#> SELECT `cyl`, COUNT(*) AS `n`
#> FROM `mtcars`
#> GROUP BY `cyl`