nycflights13 {dbplyr}	R Documentation

Database versions of the nycflights13 data

Description

These functions cache the data from the nycflights13 database in a local database, for use in examples and vignettes. Indexes are created to making joining tables on natural keys efficient.

Usage


nycflights13_sqlite(path = NULL)

nycflights13_postgres(dbname = "nycflights13", ...)

has_nycflights13(type = c("sqlite", "postgresql"), ...)

copy_nycflights13(con, ...)

Arguments

path
location of SQLite database file

dbname, ...
Arguments passed on to src_postgres()

[Package dbplyr version 2.1.1 Index]