Metadata about a remote table

remote_name() gives the unescaped name of the remote table, or NULL if it is a query (created by sql()) or already escape (created by ident_q()). remote_table() gives the remote table or the query. remote_query() gives the text of the query, and remote_query_plan() the query plan (as computed by the remote database). remote_src() and remote_con() give the dplyr source and DBI connection respectively.

Usage

remote_name(x, null_if_local = TRUE)

remote_table(x, null_if_local = TRUE)

remote_src(x)

remote_con(x)

remote_query(x, cte = FALSE, sql_options = NULL)

remote_query_plan(x, ...)

Arguments

x

Remote table, currently must be a tbl_sql.

null_if_local

Return NULL if the remote table is created via tbl_lazy() or lazy_frame()?

cte

Use the sql_options argument instead.

sql_options

SQL rendering options generated by sql_options().

...

Additional arguments passed on to methods.

Value

A string, or NULL if not a remote table, or not applicable. For example, computed queries do not have a "name".

Examples

mf <- memdb_frame(x = 1:5, y = 5:1, .name = "blorp")
remote_name(mf)
#> [1] "blorp"
remote_src(mf)
#> <S4 class ‘SQLiteConnection’ [package “RSQLite”] with 8 slots>NULL
remote_con(mf)
#> <SQLiteConnection>
#>   Path: :memory:
#>   Extensions: TRUE
remote_query(mf)
#> <SQL> SELECT *
#> FROM `blorp`

mf2 <- dplyr::filter(mf, x > 3)
remote_name(mf2)
#> NULL
remote_src(mf2)
#> <S4 class ‘SQLiteConnection’ [package “RSQLite”] with 8 slots>NULL
remote_con(mf2)
#> <SQLiteConnection>
#>   Path: :memory:
#>   Extensions: TRUE
remote_query(mf2)
#> <SQL> SELECT `blorp`.*
#> FROM `blorp`
#> WHERE (`x` > 3.0)