Metadata about a remote table

remote_name() gives the unescaped name of the remote table, or NULL if it is a query (created by sql()) or already escape (created by ident_q()). remote_table() gives the remote table or the query. remote_query() gives the text of the query, and remote_query_plan() the query plan (as computed by the remote database). remote_src() and remote_con() give the dplyr source and DBI connection respectively.

Usage

remote_name(x, null_if_local = TRUE)

remote_table(x, null_if_local = TRUE)

remote_src(x)

remote_con(x)

remote_query(x, cte = FALSE, sql_options = NULL)

remote_query_plan(x, ...)

Arguments

x

Remote table, currently must be a tbl_sql.

null_if_local

Return NULL if the remote table is created via tbl_lazy() or lazy_frame()?

cte

Use the render_otions argument instead.

sql_options

SQL rendering options generated by sql_options().

...

Additional arguments passed on to methods.

Value

The value, or NULL if not remote table, or not applicable. For example, computed queries do not have a "name"

Examples

mf <- memdb_frame(x = 1:5, y = 5:1, .name = "blorp")
remote_name(mf)
#> [1] "blorp"
remote_src(mf)
#> src:  sqlite 3.44.2 [:memory:]
#> tbls: blorp, dbplyr_1Cml0eFB3w, dbplyr_9LW8T4tdtb, dbplyr_9Qp2zFZpg2,
#>   dbplyr_BF1u2mpETM, dbplyr_BXbMSEFW4V, dbplyr_IMJcfubw23,
#>   dbplyr_KIirzAEU5F, dbplyr_KKmSYTlZoL, dbplyr_LZpaEplUMP,
#>   dbplyr_M3jF2vDWNx, dbplyr_N5fr153Ttz, dbplyr_TegCrj65Pg,
#>   dbplyr_VARKBRGD69, dbplyr_VqTWF7guLJ, dbplyr_hMmfTwv6Uu,
#>   dbplyr_k6wLO51oEX, dbplyr_lubn6mIqWv, dbplyr_mplkloYfhk,
#>   dbplyr_p6MFo4KAwO, dbplyr_rJy2v5ieEs, dbplyr_rKB7dOjyou,
#>   dbplyr_v8XMpd4o9O, df, dplyr::band_instruments, dplyr::band_members,
#>   mtcars, sqlite_stat1, sqlite_stat4, squirrels
remote_con(mf)
#> <SQLiteConnection>
#>   Path: :memory:
#>   Extensions: TRUE
remote_query(mf)
#> <SQL> SELECT *
#> FROM `blorp`

mf2 <- dplyr::filter(mf, x > 3)
remote_name(mf2)
#> NULL
remote_src(mf2)
#> src:  sqlite 3.44.2 [:memory:]
#> tbls: blorp, dbplyr_1Cml0eFB3w, dbplyr_9LW8T4tdtb, dbplyr_9Qp2zFZpg2,
#>   dbplyr_BF1u2mpETM, dbplyr_BXbMSEFW4V, dbplyr_IMJcfubw23,
#>   dbplyr_KIirzAEU5F, dbplyr_KKmSYTlZoL, dbplyr_LZpaEplUMP,
#>   dbplyr_M3jF2vDWNx, dbplyr_N5fr153Ttz, dbplyr_TegCrj65Pg,
#>   dbplyr_VARKBRGD69, dbplyr_VqTWF7guLJ, dbplyr_hMmfTwv6Uu,
#>   dbplyr_k6wLO51oEX, dbplyr_lubn6mIqWv, dbplyr_mplkloYfhk,
#>   dbplyr_p6MFo4KAwO, dbplyr_rJy2v5ieEs, dbplyr_rKB7dOjyou,
#>   dbplyr_v8XMpd4o9O, df, dplyr::band_instruments, dplyr::band_members,
#>   mtcars, sqlite_stat1, sqlite_stat4, squirrels
remote_con(mf2)
#> <SQLiteConnection>
#>   Path: :memory:
#>   Extensions: TRUE
remote_query(mf2)
#> <SQL> SELECT `blorp`.*
#> FROM `blorp`
#> WHERE (`x` > 3.0)