src_dbi {dbplyr}	R Documentation

Database src

Description

Since can generate a tbl() directly from a DBI connection we no longer recommend using src_dbi().

Usage

src_dbi(con, auto_disconnect = FALSE)

Arguments

con
An object that inherits from DBI::DBIConnection, typically generated by DBI::dbConnect

auto_disconnect
Should the connection be automatically closed when the src is deleted? Set to TRUE if you initialize the connection the call to src_dbi(). Pass NA to auto-disconnect but print a message when this happens.

Value

An S3 object with class src_dbi, src_sql, src.

[Package dbplyr version 1.4.4 Index]