tbl_sql {dbplyr}	R Documentation

Create an SQL tbl (abstract)

Description

Generally, you should no longer need to provide a custom tbl() method. The default tbl.DBIConnect method should work in most cases.

Usage


tbl_sql(subclass, src, from, ..., vars = NULL)

Arguments

subclass
name of subclass

...
needed for agreement with generic. Not otherwise used.

vars
Provide column names as a character vector to avoid retrieving them from the database. Mainly useful for better performance when creating multiple tbl objects.

[Package dbplyr version 2.1.1 Index]