Create an SQL tbl (abstract)

This function is no longer needed, since backends don't need to create their own custom classes. Instead, rely on the default tbl.DBIConnection() method.

Usage

tbl_sql(subclass, src, from, ..., vars = NULL, check_from = deprecated())

Arguments

subclass

name of subclass

src

A DBIConnection object produced by DBI::dbConnect().

from

Either a table identifier or a literal sql() string.

Use a string to identify a table in the current schema/catalog. We recommend using I() to identify a table outside the default catalog or schema, e.g. I("schema.table") or I("catalog.schema.table"). You can also use in_schema()/in_catalog() or DBI::Id().

...

needed for agreement with generic. Not otherwise used.

vars

Optionally, provide a character vector of column names. If not supplied, will be retrieved from the database by running a simple query. Mainly useful for better performance when creating many tbls.

check_from