testing {dbplyr}	R Documentation

Infrastructure for testing dplyr

Description

Register testing sources, then use test_load() to load an existing data frame into each source. To create a new table in each source, use test_frame().

Usage

test_register_src(name, src)

test_register_con(name, ...)

src_test(name)

test_load(
  df,
  name = unique_table_name(),
  srcs = test_srcs$get(),
  ignore = character()
)

test_frame(..., srcs = test_srcs$get(), ignore = character())

Examples

## Not run: 
test_register_src("sqlite", {
  DBI::dbConnect(RSQLite::SQLite(), ":memory:", create = TRUE)
})

test_frame(x = 1:3, y = 3:1)
test_load(mtcars)

## End(Not run)

[Package dbplyr version 2.3.4 Index]